Added Analysis folder to library, and added plotting of single spin up to this
