+++ scripts/plot_spin_up.py
@@ -8,97 +8,7 @@ import warnings
 import os
 warnings.filterwarnings("ignore")
 
-import scipy.interpolate as inp
-from control_parameters import parameters
-locals().update(parameters)
-
-def my_interp2d(f, rad, radnew):
-    r = rad
-    rnew = radnew
-    fnew = np.zeros_like(f)
-    for i in range(f.shape[0]):
-        val = f[i, :]
-        tckp = inp.splrep(r, val)
-        fnew[i, :] = inp.splev(rnew, tckp)
-
-    return fnew
-
-def plot_stream(r,theta,vr_n,vtheta_n,density,label=None,clim=[0,0]):
-
-    rad = np.linspace(r[-1], r[0], len(r))
-    theta = np.linspace(0,np.pi,len(theta))
-
-    rr, ttheta = np.meshgrid(rad, theta)
-    plt.figure()
-    fig,ax = plt.subplots(1,1,figsize=(6,6),subplot_kw={'projection': 'polar'})
-
-    un = vr_n[:,::-1]
-    vn = vtheta_n[:,::-1]/rr[:,::-1]
-
-    un = my_interp2d(un, r[::-1], rad)
-    vn = my_interp2d(vn, r[::-1], rad)
-
-    ax.set_theta_zero_location('N')
-    ax.set_theta_direction(-1)
-    ax.set_rorigin(0)
-    ax.set_ylim(r.min(),r.max())
-    ax.set_thetamin(0)
-    ax.set_thetamax(180)
-    ax.grid(False)
-    ax.set_xticks([])
-    ax.set_yticks([])
-
-    strm = ax.streamplot(ttheta.T,rr.T,vn.T,un.T,color='#d95f02',density=density,
-                         broken_streamlines=True,linewidth=1)
-
-    fig.tight_layout()
-
-def coords_angular(path: str) -> np.ndarray | np.ndarray:
-    data = h5py.File(path, mode='r')
-    u_n_phi = data['tasks']['u_n_phi']
-    r = u_n_phi.dims[3][0][:].ravel()
-    theta = u_n_phi.dims[2][0][:].ravel()
-    return r, theta
-
-def get_angular(rs: np.ndarray, thetas: np.ndarray, u_phi: np.ndarray) -> np.ndarray:
-    omega=np.zeros((len(thetas),len(rs)))
-    for i in range(len(rs)):
-        omega[:,i]=u_phi[:,i]/(rs[i]*np.sin(thetas)[:])
-    return omega
-
-def plot_angular(path: str, t: int, ax: matplotlib.projections.polar.PolarAxes, rotating: bool) -> None:
-    
-    '''
-    Takes an output of viscous_sphere.py and plots the angular velocity.
-
-    :param path: Path to an AZ_avg_s*.h5 file.
-    :param t: Integer used to select the time plotted.
-    :param ax: Pre-defined matplotlib polar axis on which to plot the data.
-    '''
-    data = h5py.File(path, mode='r')
-    u_n_phi = data['tasks']['u_n_phi'][t,-1,:,:]
-    r, theta = coords_angular(path)
-    u_n_background = np.zeros_like(u_n_phi)
-    if not rotating:
-       for i in range(len(r)):
-        u_n_background[:,i]= Omega_Init*(r[i]*np.sin(theta)[:])
-    
-    du_n_phi = u_n_phi - u_n_background
-    omega=get_angular(r,theta,du_n_phi)
-
-    time = np.array(data['scales/sim_time'])
-    r_m, theta_m = np.meshgrid(r, theta)
-    ax.pcolormesh(theta_m,r_m,omega,clim=(0,Delta_Omega),cmap='RdBu_r',edgecolors='face')
-    ax.set_theta_zero_location('N')
-    ax.set_theta_direction(-1)
-    ax.set_rorigin(0)
-    ax.set_ylim(r.min(),r.max())
-    ax.set_thetamin(0)
-    ax.set_thetamax(180)
-    ax.grid(False)
-    ax.set_xticks([])
-    ax.set_yticks([])
-    ax.set_title(r'$t =$'+str(time[t])[:4])
+from gains.Analysis.Analyse_spin_up import *
 
 '''
 Plots angular velocity at different times.
@@ -163,7 +73,7 @@ plt.ylabel("$\Delta \Omega$")
 plt.show()
 #plt.savefig("spin_up_time_equator.png", dpi=300)
 
-
+'''
 num_files = len(path_list)
 count = 0
 for i in range(0,num_files):
@@ -177,3 +87,4 @@ for i in range(0,num_files):
         count = count+1
         if count % 20 == 0:
             print("saved frame %04d.png" % count)
+'''
