+++ src/gains/Analysis/__init__.py
@@ -1,0 +1,1 @@
+"""Holds scripts for plotting and analysing simulation outputs"""
